igop: add sub cmd export
