+++ cmd/internal/export/exportpkg.go
@@ -1,4 +1,4 @@
-package main
+package export
 
 import (
 	"errors"
