+++ cmd/internal/export/loader.go
@@ -1,4 +1,4 @@
-package main
+package export
 
 import (
 	"fmt"
