+++ cmd/internal/export/typeparam_go117.go
@@ -1,7 +1,7 @@
 //go:build !go1.18
 // +build !go1.18
 
-package main
+package export
 
 import "go/types"
